feat: Refactor epoch_service with modular types and add commitment replay support
