+++ crates/actors/src/epoch_service/mod.rs
@@ -1,0 +1,13 @@
+pub mod commitment_state;
+pub mod epoch_replay_data;
+pub mod epoch_service;
+pub mod epoch_service_config;
+pub mod epoch_service_messages;
+pub mod partition_assignments;
+
+pub use commitment_state::*;
+pub use epoch_replay_data::*;
+pub use epoch_service::*;
+pub use epoch_service_config::*;
+pub use epoch_service_messages::*;
+pub use partition_assignments::*;
